move apps models below resources
