+++ front/pages/api/w/[wId]/apps/[aId]/clone.ts
@@ -8,7 +8,7 @@ import { getDatasets } from "@app/lib/api/datasets";
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/wrappers";
 import { Authenticator } from "@app/lib/auth";
 import type { SessionWithUser } from "@app/lib/iam/provider";
-import { App, Clone, Dataset } from "@app/lib/models/apps";
+import { App, Clone, Dataset } from "@app/lib/resources/storage/models/apps";
 import { generateLegacyModelSId } from "@app/lib/resources/string_ids";
 import logger from "@app/logger/logger";
 import { apiError } from "@app/logger/withlogging";
