+++ front/pages/api/w/[wId]/apps/[aId]/datasets/[name]/index.ts
@@ -10,7 +10,7 @@ import { getDatasetHash } from "@app/lib/api/datasets";
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/wrappers";
 import type { Authenticator } from "@app/lib/auth";
 import { checkDatasetData } from "@app/lib/datasets";
-import { Dataset } from "@app/lib/models/apps";
+import { Dataset } from "@app/lib/resources/storage/models/apps";
 import logger from "@app/logger/logger";
 import { apiError } from "@app/logger/withlogging";
