+++ front/pages/api/w/[wId]/apps/[aId]/runs/index.ts
@@ -9,8 +9,8 @@ import { getDustAppSecrets } from "@app/lib/api/dust_app_secrets";
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/wrappers";
 import { Authenticator } from "@app/lib/auth";
 import type { SessionWithUser } from "@app/lib/iam/provider";
-import { App, Provider } from "@app/lib/models/apps";
 import { RunResource } from "@app/lib/resources/run_resource";
+import { App, Provider } from "@app/lib/resources/storage/models/apps";
 import { dumpSpecification } from "@app/lib/specification";
 import logger from "@app/logger/logger";
 import { apiError } from "@app/logger/withlogging";
